+++ LaserTag.X/IRTransmitter.h
@@ -2,12 +2,13 @@
 #define	IRTRANSMITTER_H
 
 #include <stdbool.h>
+#include <stdint.h>
 
 void initializeTransmitter(void);
 void handleTransmissionTimingInterrupt(void);
 
 // Asynchronously transmit the given bits over IR. If a transmission is in
 // progress, returns false and does nothing. Otherwise returns true
-bool transmitAsync(unsigned int data);
+bool transmitAsync(uint16_t data);
 
 #endif	/* IRTRANSMITTER_H */
