+++ LaserTag.X/LEDDisplay.h
@@ -1,8 +1,10 @@
 #ifndef LEDDISPLAY_H
 #define	LEDDISPLAY_H
 
+#include <stdint.h>
+
 void initializeLEDDisplay(void);
 // LSB is the first LED. 1 is on, 0 is off.
-void setLEDDisplay(unsigned int bits);
+void setLEDDisplay(uint16_t bits);
 
 #endif	/* LEDDISPLAY_H */
